# docs/swagger.yaml
basePath: /
definitions:
  controller.JamKerja:
    properties:
      durasi:
        type: integer
      gmt:
        type: integer
      hari:
        items:
          type: string
        type: array
      jam_keluar:
        type: string
      jam_masuk:
        type: string
      piket_tim:
        type: string
      shift:
        type: integer
    type: object
  controller.Karyawan:
    properties:
      _id:
        type: string
      hari_kerja:
        items:
          type: string
        type: array
      jabatan:
        type: string
      jam_kerja:
        items:
          $ref: '#/definitions/controller.JamKerja'
        type: array
      nama:
        type: string
      phone_number:
        type: string
    type: object
  controller.Presensi:
    properties:
      _id:
        type: string
      biodata:
        $ref: '#/definitions/controller.Karyawan'
      checkin:
        description: Datetime     primitive.DateTime `bson:"datetime,omitempty" json:"datetime,omitempty"`
        type: string
      latitude:
        type: number
      location:
        type: string
      longitude:
        type: number
      phone_number:
        type: string
    type: object
host: agita.herokuapp.com
info:
  contact:
    email: 1214029@std.ulbi.ac.id
    name: API Support
    url: https://github.com/agitanurfd
  description: This is a sample server
  title: Swagger Example API
  version: "1.0"
paths:
  /ins:
    post:
      consumes:
      - application/json
      description: Input data presensi.
      parameters:
      - description: Payload Body [RAW]
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.Presensi'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Insert data presensi.
      tags:
      - Presensi
  /presensi:
    get:
      consumes:
      - application/json
      description: Mengambil semua data presensi.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
      summary: Get All Data Presensi.
      tags:
      - Presensi
  /presensi/{id}:
    get:
      consumes:
      - application/json
      description: Ambil per ID data presensi.
      parameters:
      - description: Masukan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Presensi'
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      summary: Get By ID Data Presensi.
      tags:
      - Presensi
schemes:
- http
- https
swagger: "2.0"